+++ packages/ui/src/stories/icon/AllIcon.stories.ts
@@ -1,0 +1,14 @@
+import allIcons from './allIcons.svelte';
+import type { Meta, StoryObj } from '@storybook/svelte';
+
+const meta = {
+	title: 'Elements / Icon / All icons',
+	component: allIcons
+} satisfies Meta<allIcons>;
+
+export default meta;
+type Story = StoryObj<typeof meta>;
+
+export const IconStory: Story = {
+	name: 'All icons'
+};
